refactor(serialize): break up serializers into separate files
